+++ library/js/scripts.js
@@ -96,6 +96,20 @@ $(document).ready(function($) {
 	}
 	adjustFixedHeader();
  	
+	
+	// apply fancybox to galleries
+	$('.gallery').each( function( index ) {
+		$(this).find('dl').each( function() {
+			// enable captions in fancybox
+			var dd = $(this).find('dd');
+			if ( dd ) {
+				$(this).find('a').attr('title', dd.text());
+			}
+			// add all links in this caption to a group and call fancybox
+			$(this).find('a').attr('data-fancybox-group', index).fancybox();
+		});
+	});
+	
 }); /* end of as page load scripts */